+++ workers/ohsome_quality_analyst/api/api.py
@@ -280,47 +280,47 @@ async def get_available_regions(asGeoJSON: bool = False):
         return response
 
 
-@app.get("/indicatorLayerCombinations")
-async def list_indicator_layer_combinations():
-    """List names of available indicator-layer-combinations."""
+@app.get("/indicator-layer-combinations")
+async def get_indicator_layer_combinations():
+    """Get names of available indicator-layer combinations."""
     response = empty_api_response()
     response["result"] = INDICATOR_LAYER
     return response
 
 
-@app.get("/indicatorNames")
-async def list_indicators():
-    """List names of available indicators."""
+@app.get("/indicators")
+async def indicator_names():
+    """Get names of available indicators."""
     response = empty_api_response()
     response["result"] = get_indicator_names()
     return response
 
 
-@app.get("/datasetNames")
-async def list_datasets():
-    """List names of available datasets."""
+@app.get("/datasets")
+async def dataset_names():
+    """Get names of available datasets."""
     response = empty_api_response()
     response["result"] = get_dataset_names_api()
     return response
 
 
-@app.get("/layerKeys")
-async def list_layers():
-    """List names of available layers."""
+@app.get("/layers")
+async def layer_names():
+    """Get names of available layers."""
     response = empty_api_response()
     response["result"] = get_layer_keys()
     return response
 
 
-@app.get("/reportNames")
-async def list_reports():
-    """List names of available reports."""
+@app.get("/reports")
+async def report_names():
+    """Get names of available reports."""
     response = empty_api_response()
     response["result"] = get_report_names()
     return response
 
 
-@app.get("/fidFields")
+@app.get("/fid-fields")
 async def list_fid_fields():
     """List available fid fields for each dataset."""
     response = empty_api_response()